Compilation: config.sh: Create configuration options for cc toolchain.
The system is going to need to be able to invoke the custom cross
compiling toolchain directly from the directories it is being compiled
from. To do this we'll need to make sure our cross compiler is present
in the host systems environmental variables. config.sh will export
the needed path variables when executed.

INFO: Parts of this file were excerpted from osdev.com

Signed-off-by: Isaac Assegai <isaac.a.travers@gmail.com>